requirements.txt
python-dotenv
pygame
pyinstaller
